+++ tests/countsToBinomial.R
@@ -1,6 +1,0 @@
-library(BradleyTerry2)
-data(citations, package = "BradleyTerry2")
-
-## Convert frequencies to success/failure data
-results <- countsToBinomial(citations)
-results